+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Mini dashboard: Remove javascript, fix margins, add link to main dashboard
 * Fix missing aquos_generate_signature function
 * Prevent non-admins from seeing  dashboard calculations
+* New setting: Blocked before date
 
 ### 1.4.4: May 27th, 2021
 * Dialog Insight: Make birthdate value optional when sending to web service